Added support for dismissing the color help menu with keyboard input.
The color help menu can be displayed with keyboard and joypad input but can only be dismissed with input on the joypad. Added support for dismissing the color help menu with keyboard input. Otherwise users with keyboard only control can become stuck in the color help menu.

Added this to the credits menu, even though it appears to be currently unused.